16zadanie_06.05.2026

+++ app/Http/Controllers/ReportController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Report;
 use App\Models\Status;
+use Illuminate\Support\Facades\Auth;
 
 class ReportController extends Controller
 {
@@ -22,11 +23,13 @@ class ReportController extends Controller
     
     if($validate){
         $reports = Report::where('status_id', $status)
+            ->where('user_id', Auth::user()->id)
             ->orderBy('created_at', $sort)
             ->paginate(8);
     } else {
-        $reports = Report::orderBy('created_at', $sort)
-            ->simplepaginate(8); 
+        $reports = Report::where('user_id', Auth::user()->id)
+            ->orderBy('created_at', $sort)
+            ->paginate(8);
     }
 
     $statuses = Status::all();
@@ -35,9 +38,15 @@ class ReportController extends Controller
 }
 
     public function destroy(Report $report){
-        $report->delete();
-        return redirect()->back();
-    }
+        if (Auth::user()->id === $report->user_id) {
+            $report->delete();
+            return redirect()->back();
+        } 
+        else {
+            abort(403, 'У вас нет прав на редактирование этой записи');
+            }
+        }
+
 
     public function store(Request $request, Report $report){
         $data = $request->validate([
@@ -45,15 +54,29 @@ class ReportController extends Controller
             'description' => 'string',        
         ]);
 
+        $data['user_id'] = Auth::user()->id;
+        $data['status_id'] = 1;
+
         $report->create($data);
         return redirect()->back();
     } 
 
     public function edit(Report $report){
+        if (Auth::user()->id === $report->user_id) {
         return view('reports.edit', compact('report'));
+    } 
+    else {
+        abort(403, 'У вас нет прав на редактирование этой записи');
+        }
     }
 
     public function update(Request $request, Report $report){
+        if (Auth::user()->id === $report->user_id) {
+            return view('reports.edit', compact('report'));
+        } 
+        else {
+            abort(403, 'У вас нет прав на редактирование этой записи');
+            }
         $data = $request->validate([
             'number' => 'string | required',
             'description' => 'string | required',
